requirements.txt
google-api-python-client
Flask
oracledb
google-cloud-bigquery
pandas
fiona
pygeos
db-dtypes